% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{clean_2_offsetMatrix}
\alias{clean_2_offsetMatrix}
\title{Generates the offset matrix.}
\usage{
clean_2_offsetMatrix(clean_data)
}
\arguments{
\item{clean_data}{path to the .txt file.}
}
\value{
an offset matrix.
}
\description{
Generates the offset matrix.
}
