% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman2imzML.R
\name{convert}
\alias{convert}
\title{convert: converts .txt files from Renishaw's raman instruments and transforms it into .imzML files}
\usage{
convert(txt_path, imzML_path)
}
\arguments{
\item{txt_path}{path to the .txt file.}

\item{imzML_path}{path to the folder where the .imzML file is going to be stored.}
}
\value{
TRUE if everything is alright.
}
\description{
convert: converts .txt files from Renishaw's raman instruments and transforms it into .imzML files
}
