% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFIVE.R
\name{FIVE_convert}
\alias{FIVE_convert}
\title{Converter from FIVE(WiTec) text format to imzML.}
\usage{
FIVE_convert(
  info_txt_path,
  table_txt_path,
  spectrum_txt_path = NULL,
  imzML_path = NULL,
  file_name = NULL
)
}
\arguments{
\item{info_txt_path}{path to the text file contaiting the information of the adquisition.}

\item{table_txt_path}{path to the text file contaiting all spectra as a table.}

\item{spectrum_txt_path}{(optional) path to the text file contaiting one independent spectrum to correct the raman shift axis in the table text file.}

\item{imzML_path}{(optional) path to the folder where the imzML file is going to be stored. By default the same as the table text file.}

\item{file_name}{(optional) name of the imzML file. By default the same as the table text file.}
}
\value{
complete path of the imzML file.
}
\description{
Converts a text files exported using FIVE 5.1 from WiTec raman instruments and transforms it into an imzML file. The name
of the imzML file is going to be the same as the table txt file. Only FIVE 5.1 text files have been tested.
}
