% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFIVE.R
\name{five_raw_2_clean}
\alias{five_raw_2_clean}
\title{Gives clean format to the raw FIVE text format.}
\usage{
five_raw_2_clean(raw_data, info, spect)
}
\arguments{
\item{raw_five}{data frame containing the raw data from the txt file.}
}
\value{
clean data.
}
\description{
Gives clean format to the raw FIVE text format.
}
