% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Raman2imzML}
\alias{Raman2imzML}
\title{Raman2imzML}
\description{
R package to convert raman data in .txt format exported from WiRe 5.2(Renishaw) and FIVE 5.1(WiTec) into the imzML data format. This package is intended to handle ONLY IMAGING DATA. This means that point sampling or lines are not supported.
}
\author{
Lluc Semente Fernandez
}
