% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman2imzML.R
\name{raw_2_clean}
\alias{raw_2_clean}
\title{raw_2_clean}
\usage{
raw_2_clean(raw_data)
}
\arguments{
\item{raw_data}{data frame containing the raw data from the txt file.}
}
\value{
rearranged data to fit the processing pipeline.
}
\description{
raw_2_clean
}
