% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWiRe.R
\name{WiRe_convert}
\alias{WiRe_convert}
\title{Converter from WiRe(Renishaw) text format to imzML.}
\usage{
WiRe_convert(txt_path, imzML_path = NULL, file_name = NULL)
}
\arguments{
\item{txt_path}{path to the txt file.}

\item{imzML_path}{path to the folder where the imzML file is going to be stored. By default the same as the text file.}

\item{file_name}{name of the imzML file. By default the same as txt file.}
}
\value{
complete path of the imzML file.
}
\description{
Converts a txt files exported using WiRe 5.2 from Renishaw raman instruments and transforms it into an imzML file. The name
of the imzML file is going to be the same as the txt file. Only WiRe 5.2 txt files have been tested.
}
