% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWiRe.R
\name{wire_raw_2_clean}
\alias{wire_raw_2_clean}
\title{Gives clean format to the raw WiRe text format.}
\usage{
wire_raw_2_clean(raw_data)
}
\arguments{
\item{raw_data}{data frame containing the raw data from the txt file.}
}
\value{
clean data.
}
\description{
Gives clean format to the raw WiRe text format.
}
