% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman2imzML.R
\name{write_ibd}
\alias{write_ibd}
\title{write_ibd}
\usage{
write_ibd(imzML_path, file_name, clean_data, uuid)
}
\arguments{
\item{imzML_path}{raw_data.}

\item{file_name}{raw_data.}

\item{clean_data}{raw_data.}

\item{uuid}{raw_data.}
}
\value{
clean data.
}
\description{
write_ibd
}
