% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman2imzML.R
\name{clean_2_offsetMatrix}
\alias{clean_2_offsetMatrix}
\title{clean_2_offsetMatrix}
\usage{
clean_2_offsetMatrix(clean_data)
}
\arguments{
\item{clean_data}{path to the .txt file.}
}
\value{
TRUE if everything is alright.
}
\description{
clean_2_offsetMatrix
}
